Refactoring, renaming, bug fixes and new features
Renamed to Droidule, refactored to a package based hierargy.

Fixed a bug where people can't scroll upwars on the attendee selection screen.
Changed the star icon to a home because I think it's more fitting.

Added a super cool theme engine with 3 supported default themes.
Added a feature to show next weeks schedule during the weekends, instead of the current week, as that's kinda useless.
Added a feature that will automatically select the first chosen schedule to be the homescreen (Be starred) so it even if people don't understand the mechanic it will still work.